requirements.txt
Django==3.2.8
djangorestframework==3.12.4
python-dotenv==0.19.0
mysqlclient==2.0.3
pillow==8.2.0
celery==5.1.2
django-celery-results==2.2.0
influxdb-client==1.21.0
git+git://github.com/save-the-computer/collector@0.1.1